Add GSCore class
This class takes care of the basic communication with the module. It
handles handling the various escape sequences, UART and SPI
communication, sending AT commands and interpreting responses.

The code added in this commit has the basic infrastructure in place. The code
for sending specific commands and keeping track of clients/servers will be
added in separate classes.  There are also a lot of unhandled corner cases
still.