Support data ready pin and improve polling without it
The Gainspan module supports a data ready to indicate wether there is
SPI data waiting. The library now supports checking this pin to quickly
determine if there is data waiting or not.

Additionally, this commit improves the polling behaviour previously
used. The gainspan module does something weird: In idle mode it fills up
its SPI buffer with idle bytes, which don't get removed when data
becomes available. This means we'll have to read out all of those bytes
before we get at the data and that we'll aslo have to read all of them
(plus one) to see if there was any data in the first place...