Fix response handling for bulk data frames
When sending a bulk data frame, the module sends the <ESC>O or <ESC>F
reponse after the cid is received, instead of after the entire packet as
previously expected. For short packets, this did you cause problems
(because there were 63 idle bytes before the actual response). However,
for longer packets, the response would be received while still sending
data, causing it to be discarded and the later readDataResponse() to
lock up. Also, in case of a failure, the module would interpret the data
sent as commands instead of data, potentially causing arbitrary
behaviour and problems.

Now, the code sends the first half of the escape sequence, waits for a
response, then continues to send the data. This increases the time it
takes for sending data (especially due to the 63 idle bytes problem),
but there seems to be no other reliable way to handle this.