Fix infinite loop in setAutoConnectClient
Before, the IPAddress object passed was converted to a string in a
uint8_t[] and another version of setAutoConnectClient was called passing
in the buffer. However, that other versions accepted a char* instead of
a uint8_t* and the compiler apparently considered casting the uint8_t[]
to an IPAddress easier than casting it to char*, causing the method to
call itself repeatedly.