Don't return true on error in readDataResponse()
This looks like a copy-paste error that caused the library to send the
data even when the gainspan module returned a failure for the data
header. This in turn would cause the module to interpret the data as a
command (and usually return an INVALID INPUT error, but perhaps other
random breakage is also possible).