Update tail_frame.length in readData
When reading data from the buffer, this value was not updated. Depending
on the timing, this could cause non-date bytes to be interpreted as data
(and, since this could cause escape bytes to be missed, subsequently
could cause data to be dropped). It seems this did not often occur,
because most often data was not read through the buffer but directly.