Advance the tail pointer in GSCore::loadFrameHeader
Before, the frame header would be loaded from the buffer but the pointer
was not advanced. This caused the frame header to be read again as data,
and then some data at the end would be read as the next frame header.
This frame header would contain garbage, so nobody would read the frame
(because the cid did not match), eventually causing rx_data overflow:

	rx_data is full, dropped byte for cid 34