Intitialize frame length in GSUdpServer constructor
When a GSUdpServer object is intialized on the stack (like in the
GSUdpServer example), its variables will not be zero-initialized. Having
a random value for the packet length will cause all kinds of weirdness.

Initializing just the packet length to 0 should be enough, all other
values are unused when length is 0.